Nonlocal material model working (g_nonlocal.c).
Fixed up some of the boundary and loading conditions to make this
happen, but there are still some questions as to what is supposed to
be solved for the nonlocal part. Currently using a mass-weighted
approach, explicit constitutive update. Also changed the reader
to apply an initial stress state if needed.