Changed default colormap for csv plotting.
I did this so I could have a more perceptually uniform map instead
of the 'jet' default. The 'summer' map also seems to work pretty
well, but I'm not sure which one I like better. Need to read up
on types of colorblindness to see if this is a good idea.